+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.4.0 - 2021-04-04
+- Fixed unsupported fields able to be selected in the search setup
+- Fixed category sort options selection.
+
 ## 1.3.2 - 2021-04-04
 - Added searchType var to EVENT_ELEMENT_QUERY eventAdded searchType var to EVENT_ELEMENT_QUERY event.
 - Fixed category element search type.